// src/app/core/models/book.model.ts
export interface Book {
    $id: string;
    titulo: string;
    autor: string;
    genero: string;
    añoPublicacion: number;
    paginas: number;
    editorial: string;
    issn: string;
    idioma: string;
    fechaPublicacion: string;
    estado: string;
    precio: number;
    image: string;
  }